# apiDoc/swagger.json
{
  "basePath": "/",
  "consumes": ["application/json"],
  "definitions": {
    "TokenAPI": {
      "properties": {
        "token": {
          "type": "string"
        }
      },
      "required": ["token"],
      "type": "object",
      "additionalProperties": false
    },
    "ErrorValidacion": {
      "properties": {
        "propertyPath": {
          "type": "string"
        },
        "code": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      },
      "required": ["propertyPath", "code", "message"],
      "type": "object",
      "additionalProperties": false
    },
    "ErrorResponse": {
      "properties": {
        "stack": {
          "type": "string",
          "x-nullable": true
        },
        "result": {
          "type": "string",
          "enum": ["error"]
        },
        "status": {
          "type": "number",
          "format": "double"
        },
        "userMessage": {
          "type": "string"
        },
        "developerMessage": {
          "type": "string"
        },
        "validationErrors": {
          "items": {
            "$ref": "#/definitions/ErrorValidacion"
          },
          "type": "array"
        }
      },
      "required": ["result", "status", "userMessage", "developerMessage", "validationErrors"],
      "type": "object",
      "additionalProperties": false
    },
    "LoginRequest": {
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": ["username", "password"],
      "type": "object",
      "additionalProperties": false,
      "example": {
        "username": "ipianetti",
        "password": "1234"
      }
    },
    "UsuarioAPI": {
      "properties": {
        "id": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "role": {
          "type": "string"
        }
      },
      "required": ["id", "username", "role"],
      "type": "object",
      "additionalProperties": false
    },
    "UsuarioResponseData": {
      "properties": {
        "result": {
          "type": "string",
          "default": "ok",
          "x-nullable": true
        },
        "status": {
          "type": "number",
          "format": "double",
          "default": 200,
          "x-nullable": true
        },
        "data": {
          "$ref": "#/definitions/UsuarioAPI"
        },
        "userMessage": {
          "type": "string"
        },
        "actions": {
          "type": "string"
        }
      },
      "required": ["data", "userMessage", "actions"],
      "type": "object",
      "additionalProperties": false
    },
    "ChangePasswordRequest": {
      "properties": {
        "oldPassword": {
          "type": "string"
        },
        "newPassword": {
          "type": "string"
        }
      },
      "required": ["oldPassword", "newPassword"],
      "type": "object",
      "additionalProperties": false,
      "example": {
        "oldPassword": "asdf",
        "newPassword": "1234"
      }
    },
    "UsuarioResponseLista": {
      "properties": {
        "result": {
          "type": "string",
          "default": "ok",
          "x-nullable": true
        },
        "status": {
          "type": "number",
          "format": "double",
          "default": 200,
          "x-nullable": true
        },
        "data": {
          "items": {
            "$ref": "#/definitions/UsuarioAPI"
          },
          "type": "array"
        },
        "userMessage": {
          "type": "string"
        },
        "actions": {
          "type": "string"
        }
      },
      "required": ["data", "userMessage", "actions"],
      "type": "object",
      "additionalProperties": false
    },
    "UsuarioCreationRequest": {
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "role": {
          "type": "string"
        }
      },
      "required": ["username", "password", "role"],
      "type": "object",
      "additionalProperties": false
    },
    "UsuarioUpdateRequest": {
      "properties": {
        "username": {
          "type": "string",
          "x-nullable": true
        },
        "role": {
          "type": "string",
          "x-nullable": true
        }
      },
      "type": "object",
      "additionalProperties": false
    }
  },
  "info": {
    "title": "jwt-express-typeorm",
    "version": "0.0.1",
    "description": "Awesome project developed with TypeORM."
  },
  "paths": {
    "/login": {
      "post": {
        "operationId": "login",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "$ref": "#/definitions/TokenAPI"
            }
          },
          "400": {
            "description": "Username y password son requeridos",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Username o password incorrecto",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "summary": "Obtención de Token JWT",
        "tags": ["Auth"],
        "security": [],
        "parameters": [
          {
            "in": "body",
            "name": "data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ]
      }
    },
    "/changePassword": {
      "put": {
        "operationId": "changePassword",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "$ref": "#/definitions/UsuarioResponseData"
            }
          },
          "400": {
            "description": "La clave nueva y actual clave son requeridas",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "La clave actual es incorrecta",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "409": {
            "description": "Errores de validación",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "summary": "Modificar clave",
        "tags": ["Auth"],
        "security": [
          {
            "access_token": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ChangePasswordRequest"
            }
          }
        ]
      }
    },
    "/api/v1/usuarios": {
      "get": {
        "operationId": "findAllUsuarios",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "$ref": "#/definitions/UsuarioResponseLista"
            }
          }
        },
        "summary": "Obtiene una lista de usuarios",
        "tags": ["Usuario"],
        "security": [
          {
            "access_token": ["SUPER_ADMIN"]
          }
        ],
        "parameters": []
      },
      "post": {
        "operationId": "saveUsuario",
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "Usuario creado correctamente",
            "schema": {
              "$ref": "#/definitions/UsuarioResponseData"
            }
          },
          "409": {
            "description": "Errores de validación",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "summary": "Permite crear un usuario",
        "tags": ["Usuario"],
        "security": [
          {
            "access_token": ["SUPER_ADMIN"]
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UsuarioCreationRequest"
            }
          }
        ]
      }
    },
    "/api/v1/usuarios/{id}": {
      "get": {
        "operationId": "findUsuario",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "$ref": "#/definitions/UsuarioResponseData"
            }
          },
          "404": {
            "description": "No se encontró usuario con ID 123",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "summary": "Obtiene un usuario por ID",
        "tags": ["Usuario"],
        "security": [
          {
            "access_token": ["SUPER_ADMIN"]
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ]
      },
      "put": {
        "operationId": "updateUsuario",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Usuario actualizado correctamente",
            "schema": {
              "$ref": "#/definitions/UsuarioResponseData"
            }
          },
          "404": {
            "description": "No se encontró usuario con ID 123",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "409": {
            "description": "Errores de validación",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "summary": "Permite actualizar un usuario por ID",
        "tags": ["Usuario"],
        "security": [
          {
            "access_token": ["SUPER_ADMIN"]
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UsuarioUpdateRequest"
            }
          }
        ]
      },
      "delete": {
        "operationId": "deleteUsuario",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Usuario eliminado correctamente",
            "schema": {
              "$ref": "#/definitions/UsuarioResponseData"
            }
          },
          "404": {
            "description": "No se encontró usuario con ID 123",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "summary": "Permite borrar un usuario por ID",
        "tags": ["Usuario"],
        "security": [
          {
            "access_token": ["SUPER_ADMIN"]
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ]
      }
    }
  },
  "produces": ["application/json"],
  "swagger": "2.0",
  "securityDefinitions": {
    "access_token": {
      "name": "Authorization",
      "type": "apiKey",
      "in": "header",
      "description": "Access Token JWT"
    }
  }
}